+++ app/src/main/java/com/iron/espresso/di/RepositoryModule.kt
@@ -1,7 +1,6 @@
 package com.iron.espresso.di
 
 import com.iron.espresso.domain.repo.*
-import com.iron.espresso.local.model.ChatRepository
 import com.iron.espresso.local.model.ChatRepositoryImpl
 import com.iron.espresso.model.repo.*
 import dagger.Binds
