+++ app/src/main/java/com/iron/espresso/domain/repo/ChatRepository.kt
@@ -1,27 +1,18 @@
-package com.iron.espresso.local.model
+package com.iron.espresso.domain.repo
 
-import androidx.room.Dao
-import androidx.room.Delete
-import androidx.room.Insert
-import androidx.room.Query
+import com.iron.espresso.local.model.ChatEntity
 import io.reactivex.Completable
 import io.reactivex.Flowable
 import io.reactivex.Single
 
-@Dao
-interface ChatDao {
-    @Query("SELECT * FROM chat_table WHERE study_id IN (:studyId)")
+interface ChatRepository {
     fun getAll(studyId: Int): Flowable<List<ChatEntity>>
 
-    @Query("SELECT max(time_stamp) FROM chat_table WHERE study_id IN (:studyId)")
     fun getTimeStamp(studyId: Int): Single<Long>
 
-    @Insert
     fun insert(chatEntity: ChatEntity): Completable
 
-    @Insert
     fun insertAll(chatEntity: List<ChatEntity>): Completable
 
-    @Delete
     fun delete(chatEntity: ChatEntity): Completable
 }
